wip: repro case for creating a test using a session
Chronograf uses Sessions, but the existing tests use Authorizations.
Right now we are incorrectly storing a Session ID as an Authorization
ID, and that seems to make tasks hang.

This new test doesn't quite repro it yet, because it doesn't correctly
set permissions on the Session. Once that is addressed, I think this
test will repro the issue.